Change snacks file truncation
This is better for finding files with long names, but worse for grepping
in files with long names. For now it's an improvement for me, but there
will likely need to be some tweaks.

Before this method, the existing truncation would remove all of the
middle paths. If you had multiple files with the same names but in
different directories, it was really hard to tell which was which since
you couldn't see the intermediate directories.