Remove command picker and just use user commands
When using the picker, it's hard to re-run a command over and over again
if needed (e.g. want to keep re-grepping in the same directory). When
using the normal commands it's way easier to do that since you can just
rerun using @:

A lot of the more random commands have also been moved to be subcommands
of the :C command to make it easier to autocomplete them. This is
explained the command above the subcommands table definition.